+++ mayan/apps/documents/forms/document_version_page_forms.py
@@ -74,7 +74,21 @@ class DocumentVersionPageMappingFormSet(
         form=DocumentVersionPageMappingForm, extra=0
     )
 ):
-    """Combined formset"""
+    """
+    Combined formset
+    """
+    def clean(self):
+        set_of_target_page_numbers = set()
+        for form in self.forms:
+            cleaned_data_entry = form.cleaned_data
+            target_page_number = cleaned_data_entry['target_page_number']
+            if target_page_number in set_of_target_page_numbers:
+                form.add_error(
+                    field='target_page_number',
+                    error=_('Target page number can\'t be repeated.')
+                )
+            else:
+                set_of_target_page_numbers.add(target_page_number)
 
 
 class DocumentVersionPageNumberForm(forms.Form):