+++ mayan/apps/documents/urls.py
@@ -42,8 +42,9 @@ from .views.document_version_page_views import (
     DocumentVersionPageZoomOutView
 )
 from .views.document_version_views import (
-    DocumentVersionDeleteView, DocumentVersionEditView,
-    DocumentVersionListView, DocumentVersionPreviewView
+    DocumentVersionCreateView, DocumentVersionDeleteView,
+    DocumentVersionEditView, DocumentVersionListView,
+    DocumentVersionPreviewView
 )
 from .views.document_views import (
     DocumentDocumentTypeChangeView, DocumentListView, DocumentPreviewView,
@@ -226,6 +227,11 @@ urlpatterns_document_version = [
     url(
         regex=r'^documents/(?P<document_id>\d+)/versions/$',
         name='document_version_list', view=DocumentVersionListView.as_view()
+    ),
+    url(
+        regex=r'^documents/(?P<document_id>\d+)/versions/create/$',
+        name='document_version_create',
+        view=DocumentVersionCreateView.as_view()
     ),
     url(
         regex=r'^documents/versions/(?P<document_version_id>\d+)/delete/$',